FSL compatability changes.  Py2toPy3 changes also.
Renamed the jedi file to jedi.sh.  That's fine to change back if we want.

Also added a script to run on the fsl.  Currently, it's not working remotely, but you can ssh into the ssh.fsl nodes and run it and it works just fine.  Ticket 5061 is asking about that still.

Modified the imports (psycopg won't work on the fsl.) (py2 and py3 have different queue imports)